backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.3
cycler==0.12.1
dlib==20.0.0
dnspython==2.7.0
email-validator==2.3.0
face-recognition==1.3.0
face_recognition_models==0.3.0
faiss-cpu==1.12.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
safetensors==0.6.2
sentencepiece==0.2.1
sentry-sdk==2.35.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
structlog==25.4.0
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.4
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
